multisig: refactor threshold to uint32_t
(cherry picked from commit 5aec9db0eda9f7f7a940d6d3ab8191bc059e28b5)

Conflicts: main/multisig.c (three blocks) and main/multisig.h (one).  This fork
had already rewritten the record writer: the function is multisig_body_to_bytes()
rather than upstream's multisig_data_to_bytes(), and it is reached through
multisig_open_registration() and multisig_seal_body(), which upstream does not
have.  Those names and that call chain were kept.

Upstream's side was taken where the record depends on it: the parameter type
(uint32_t), the JADE_STATIC_ASSERT that MAX_ALLOWED_SIGNERS fits in a uint8_t,
and the single-byte writes.  Taking the type on its own would have been silently
wrong.  The fork's body wrote the field with memcpy(write_ptr, &threshold,
sizeof(threshold)), so a wider parameter would have written four bytes where
MULTISIG_BODY_LEN() counts one (its header is 5 * sizeof(uint8_t): variant,
sorted, threshold, keylen, num_signers), pushing every later field along and
breaking a record this fork can still read.  The reader takes a single byte
(multisig.c `output->threshold = *read_ptr`) and was not touched; register_multisig_process()
already refuses a threshold above MAX_ALLOWED_SIGNERS before the writer sees it.

Upstream's "(new to version 1)" and "(new to version 2)" comments were not taken;
this fork carries its own record version.

Measured in the emulator: a registration written by the binary from the commit
before this series is read back by the binary from the end of it, and the
reverse, with threshold 2 and two signers intact in both directions.

+++ main/multisig.h
@@ -60,7 +60,7 @@ typedef struct _multisig_data {
     uint8_t xpubs[MAX_ALLOWED_SIGNERS * BIP32_SERIALIZED_LEN];
 } multisig_data_t;
 
-WARN_UNUSED_RESULT bool multisig_body_to_bytes(script_variant_t variant, bool sorted, uint8_t threshold,
+WARN_UNUSED_RESULT bool multisig_body_to_bytes(script_variant_t variant, bool sorted, uint32_t threshold,
     const uint8_t* master_blinding_key, size_t master_blinding_key_len, const signer_t* signers, size_t num_signers,
     size_t total_num_path_elements, uint8_t* body, size_t body_len);
